+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/servers/AbstractTigerServer.java
@@ -4,7 +4,6 @@
 
 package de.gematik.test.tiger.testenvmgr.servers;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import de.gematik.rbellogger.util.RbelAnsiColors;
 import de.gematik.test.tiger.common.Ansi;
 import de.gematik.test.tiger.common.config.SourceType;
@@ -106,10 +105,12 @@ public abstract class AbstractTigerServer implements TigerEnvUpdateSender {
                 }
 
                 String[] routeParts = mapping.split(" --> ", 2);
-                testEnvMgr.getLocalTigerProxy().addRoute(TigerRoute.builder()
-                    .from(routeParts[0])
-                    .to(routeParts[1])
-                    .build());
+                testEnvMgr.getLocalTigerProxyOptional().ifPresent(
+                    proxy -> proxy.addRoute(TigerRoute.builder()
+                        .from(routeParts[0])
+                        .to(routeParts[1])
+                        .build())
+                );
 
             });
         }
@@ -156,11 +157,12 @@ public abstract class AbstractTigerServer implements TigerEnvUpdateSender {
                         "Your certificate is empty, please check your .yaml-file for " + key.getId());
                 }
                 log.info("Adding certificate {}", key.getId());
-                getTigerTestEnvMgr().getLocalTigerProxy().addKey(
+                getTigerTestEnvMgr().getLocalTigerProxyOptional().ifPresent(proxy -> proxy.addKey(
                     key.getId(),
                     KeyMgr.readCertificateFromPem("-----BEGIN CERTIFICATE-----\n"
                         + key.getPem().replace(" ", "\n")
-                        + "\n-----END CERTIFICATE-----").getPublicKey());
+                        + "\n-----END CERTIFICATE-----").getPublicKey())
+                );
             });
         getConfiguration().getPkiKeys().stream()
             .filter(key -> key.getType() == PkiType.Key)
@@ -170,11 +172,12 @@ public abstract class AbstractTigerServer implements TigerEnvUpdateSender {
                         "Your Key is empty, please check your .yaml-file for " + key.getId());
                 }
                 log.info("Adding key {}", key.getId());
-                getTigerTestEnvMgr().getLocalTigerProxy().addKey(
+                getTigerTestEnvMgr().getLocalTigerProxyOptional().ifPresent(proxy -> proxy.addKey(
                     key.getId(),
                     KeyMgr.readKeyFromPem("-----BEGIN PRIVATE KEY-----\n"
                         + key.getPem().replace(" ", "\n")
-                        + "\n-----END PRIVATE KEY-----"));
+                        + "\n-----END PRIVATE KEY-----"))
+                );
             });
     }
 
@@ -192,7 +195,9 @@ public abstract class AbstractTigerServer implements TigerEnvUpdateSender {
     }
 
     public void assertThatConfigurationIsCorrect() {
-        assertThat(serverId).withFailMessage("Server Id must not be blank!").isNotBlank();
+        if (StringUtils.isBlank(serverId)) {
+            throw new TigerTestEnvException("Server Id must not be blank!");
+        }
         assertCfgPropertySet(getConfiguration(), "type");
 
         // set default values for all types
@@ -257,15 +262,19 @@ public abstract class AbstractTigerServer implements TigerEnvUpdateSender {
     }
 
     void addRoute(TigerRoute newRoute) {
-        getTigerTestEnvMgr().getLocalTigerProxy().addRoute(newRoute);
-        routes.add(newRoute);
+        getTigerTestEnvMgr().getLocalTigerProxyOptional().ifPresent(proxy -> {
+            proxy.addRoute(newRoute);
+            routes.add(newRoute);
+        });
     }
 
     void removeAllRoutes() {
-        log.info("Removing routes for {}...", getServerId());
-        routes.stream()
-            .map(TigerRoute::getId)
-            .forEach(getTigerTestEnvMgr().getLocalTigerProxy()::removeRoute);
+        getTigerTestEnvMgr().getLocalTigerProxyOptional().ifPresent(proxy -> {
+            log.info("Removing routes for {}...", getServerId());
+            routes.stream()
+                .map(TigerRoute::getId)
+                .forEach(proxy::removeRoute);
+        });
     }
 
     public abstract void shutdown();